#252 Moved core-tests to integration-tests
