--- bb-core-tests/src/main/java/com/cognifide/qa/test/pageobjects/current/scope/CurrentScopeElementPageObject.java
@@ -1,40 +1,0 @@
-/*-
- * #%L
- * Bobcat
- * %%
- * Copyright (C) 2016 Cognifide Ltd.
- * %%
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- * #L%
- */
-package com.cognifide.qa.test.pageobjects.current.scope;
-
-import org.openqa.selenium.WebElement;
-
-import com.cognifide.qa.bb.qualifier.CurrentScope;
-import com.cognifide.qa.bb.qualifier.PageObject;
-import com.google.inject.Inject;
-
-@PageObject
-public class CurrentScopeElementPageObject {
-
-  public static final String CSS = ".element";
-
-  @Inject
-  @CurrentScope
-  private WebElement element;
-
-  public String getElementText() {
-    return element.getText();
-  }
-}
