+++ bb-integration-tests/src/main/config/report.properties
