+++ bb-integration-tests/src/main/java/com/cognifide/qa/bb/core/pageobjects/current/scope/CurrentScopeListPageObject.java
@@ -1,0 +1,43 @@
+/*-
+ * #%L
+ * Bobcat
+ * %%
+ * Copyright (C) 2016 Cognifide Ltd.
+ * %%
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * #L%
+ */
+package com.cognifide.qa.bb.core.pageobjects.current.scope;
+
+import java.util.List;
+import java.util.stream.Collectors;
+
+import org.openqa.selenium.WebElement;
+
+import com.cognifide.qa.bb.qualifier.CurrentScope;
+import com.cognifide.qa.bb.qualifier.PageObject;
+import com.google.inject.Inject;
+
+@PageObject
+public class CurrentScopeListPageObject {
+
+  public static final String CSS = ".list-element";
+
+  @Inject
+  @CurrentScope
+  private List<WebElement> elements;
+
+  public List<String> getElementsText() {
+    return elements.stream().map(WebElement::getText).collect(Collectors.toList());
+  }
+}
