+++ bb-integration-tests/src/main/java/com/cognifide/qa/bb/core/pageobjects/qualifier/findpageobject/MasterPage.java
@@ -1,0 +1,34 @@
+/*-
+ * #%L
+ * Bobcat
+ * %%
+ * Copyright (C) 2016 Cognifide Ltd.
+ * %%
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * #L%
+ */
+package com.cognifide.qa.bb.core.pageobjects.qualifier.findpageobject;
+
+import com.cognifide.qa.bb.qualifier.FindPageObject;
+import com.cognifide.qa.bb.qualifier.PageObject;
+
+@PageObject
+public class MasterPage {
+
+  @FindPageObject
+  private Drinks drinks;
+
+  public Drinks getDrinks() {
+    return drinks;
+  }
+}
