+++ bb-integration-tests/src/test/java/com/cognifide/qa/bb/proxy/ProxyControllerTest.java
