+++ bb-integration-tests/src/test/resources/html/com/cognifide/qa/bb/proxy/page.html
