+++ bb-integration-tests/src/test/resources/html/com/cognifide/qa/bb/test/expectedconditions/page.html
