+++ pom.xml
@@ -52,7 +52,7 @@ limitations under the License.
 
     <modules>
         <module>bb-core</module>
-        <module>bb-core-tests</module>
+        <module>bb-integration-tests</module>
         <module>bb-junit</module>
         <module>bb-reports</module>
         <module>bb-cumber</module>
